MIR refactoring
